app_form/requirements.txt
Flask==2.3.3
PyYAML==6.0.1
Werkzeug==2.3.7
openai==1.99.9
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.8.3 